pyproject.toml
[project]
name = "tasa-check"
version = "0.1.0"
description = "기자를 위한 타사 뉴스 모니터링 Telegram Bot"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "aiosqlite>=0.20.0",
    "cryptography>=43.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest-httpx>=0.36.0",
]
